requirements.txt
-i https://pypi.org/simple
click==8.1.3 ; python_version >= '3.7'
flask==2.2.2
gunicorn==20.1.0
importlib-metadata==5.1.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
setuptools==65.6.3 ; python_version >= '3.7'
werkzeug==2.2.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
yfinance==0.2.18
APScheduler~=3.10.1
DateTime~=5.1
python-dotenv==1.0.0